generator/elevenlabs/requirements.txt
elevenlabs>=1.8.0
python-dotenv>=1.1.0
requests>=2.31.0
